setup-user-lisp-directories centralised in user-directories.el.
User lisp directories creation and setup are now in
user-directories.el.  The corresponding code, which was repeated, was
factored out of the system provider functions into that function,
in lieu of which is called there.

The locator functions are now created each time the user calls
set-user-directory.  That means that, once a directory is set, the
corresponding locator function can be called.

A bug was also corrected in a case statement, that made adding the
directory to the path impossible.

  + setup-user-directories.el:
    - setup-user-lisp-directories is called in the default provider;
    - There is no more creation of locator functions after the
      provider is runned.
  + user-directories-linux.el:
    - setup-user-lisp-directories is called.
  + user-directories-windows-nt.el:
    - setup-user-lisp-directories is called.
  + user-directories.el:
    - cl is now required when compiling (for case);
    - erroneous pcase statement corrected to case;
    - created setup-user-lisp-directories.